feat: :sparkles: Route ComboBox now allows custom values
Restricting to only the input of available items
made the UI confusing because if they typed
a route ID but did not press down and actually
select that item, it would not be added to the
input box.

Now this problem is removed, but the user can now
enter an invalid route.

- Changed the label from "SRMP" to "Enter State Route & Milepost"